Move `query_account` helper function into ABI
Allow other crates to use the test helper function.
