+++ examples/fungible/tests/cross_chain.rs
@@ -5,12 +5,8 @@
 
 #![cfg(not(target_arch = "wasm32"))]
 
-use async_graphql::InputType;
-use fungible::{Account, AccountOwner, InitialStateBuilder, Operation};
-use linera_sdk::{
-    base::{Amount, ApplicationId},
-    test::{ActiveChain, TestValidator},
-};
+use fungible::{Account, AccountOwner, FungibleTokenAbi, InitialStateBuilder, Operation};
+use linera_sdk::{base::Amount, test::TestValidator};
 
 /// Test transferring tokens across microchains.
 ///
@@ -56,41 +52,14 @@ async fn test_cross_chain_transfer() {
         .await;
 
     assert_eq!(
-        query_account(application_id, sender_chain, sender_account).await,
+        FungibleTokenAbi::query_account(application_id, &sender_chain, sender_account).await,
         Some(initial_amount.saturating_sub(transfer_amount)),
     );
 
     receiver_chain.handle_received_effects().await;
 
     assert_eq!(
-        query_account(application_id, receiver_chain, receiver_account).await,
+        FungibleTokenAbi::query_account(application_id, &receiver_chain, receiver_account).await,
         Some(transfer_amount),
     );
-}
-
-/// Query the balance of an account owned by `account_owner` on a specific `chain`.
-async fn query_account<Abi>(
-    application_id: ApplicationId<Abi>,
-    chain: ActiveChain,
-    account_owner: AccountOwner,
-) -> Option<Amount> {
-    let query = format!(
-        "query {{ accounts(accountOwner: {} ) }}",
-        account_owner.to_value()
-    );
-
-    let value: serde_json::Value = chain.query(application_id, query).await;
-
-    let balance = value
-        .as_object()?
-        .get("data")?
-        .as_object()?
-        .get("accounts")?
-        .as_str()?;
-
-    Some(
-        balance
-            .parse()
-            .expect("Account balance cannot be parsed as a number"),
-    )
 }
